Adding CNAME resolution and related data.
This change looks at each request, and if it appears to be first party,
determine if there is a CNAME DNS entry for it.
If so, treat the newly resolved domain as any other request - ignore it if
it remains in the first party set, or analyze it as if it were
a normal third party requst.

All CNAME information is also stored in output for additional analysis.

Feature flag options have been added to selectively keep or ignore cname information.