Change parallell execution to sites instead of requests.
This changes cname resolution to that a single site processes requests in sequence, to make better use of
DNS caches. Previous commit had DNS requests in parallell, but that led to many duplicate requests, and a
low cache hit rate (~8%).

Now sites are executed in parallell, with a new parameter to set how many sites per batch. 100 seems to work OK.
this gets a cache hit rate closer to 60% and cuts down run time by a significant margin.